split table in smaller parts
