+++ features/discovery/common/DiscoveryTable.tsx
@@ -1,4 +1,4 @@
-import { getDiscoveryTableCellContent } from 'features/discovery/helpers'
+import { DiscoveryTableDataCellContent } from 'features/discovery/common/DiscoveryTableDataCellContent'
 import { DiscoveryTableRowData } from 'features/discovery/types'
 import { camelToKebab } from 'helpers/camelToKebab'
 import { useTranslation } from 'next-i18next'
@@ -34,56 +34,13 @@ export function DiscoveryTable({ rows }: { rows: DiscoveryTableRowData[] }) {
           <Box as="thead" sx={{}}>
             <Box as="tr">
               {Object.keys(rows[0]).map((key) => (
-                <Box
-                  as="th"
-                  key={key}
-                  sx={{
-                    px: '12px',
-                    pb: '20px',
-                    fontSize: 1,
-                    fontWeight: 'semiBold',
-                    color: 'neutral80',
-                    lineHeight: '10px',
-                    textAlign: 'right',
-                    '&:first-child': {
-                      textAlign: 'left',
-                    },
-                  }}
-                >
-                  {t(`discovery.table.header.${camelToKebab(key)}`)}
-                </Box>
+                <DiscoveryTableHeaderCell key={key} label={key} />
               ))}
             </Box>
           </Box>
           <Box as="tbody" sx={{ borderSpacing: '12px 0' }}>
             {rows.map((row, i) => (
-              <Box
-                as="tr"
-                key={i}
-                sx={{
-                  borderRadius: 'medium',
-                  transition: 'box-shadow 200ms',
-                  '&:hover': {
-                    boxShadow: 'buttonMenu',
-                  },
-                }}
-              >
-                {Object.keys(row).map((key) => (
-                  <Box
-                    as="td"
-                    key={`${key}-${i}`}
-                    sx={{
-                      p: '8px 12px',
-                      textAlign: 'right',
-                      '&:first-child': {
-                        textAlign: 'left',
-                      },
-                    }}
-                  >
-                    {getDiscoveryTableCellContent({ key, row })}
-                  </Box>
-                ))}
-              </Box>
+              <DiscoveryTableDataRow key={i} row={row} />
             ))}
           </Box>
         </Box>
@@ -92,6 +49,72 @@ export function DiscoveryTable({ rows }: { rows: DiscoveryTableRowData[] }) {
           {t('discovery.table.no-entries')}
         </Text>
       )}
+    </Box>
+  )
+}
+
+export function DiscoveryTableHeaderCell({ label }: { label: string }) {
+  const { t } = useTranslation()
+
+  return (
+    <Box
+      as="th"
+      sx={{
+        px: '12px',
+        pb: '20px',
+        fontSize: 1,
+        fontWeight: 'semiBold',
+        color: 'neutral80',
+        lineHeight: '10px',
+        textAlign: 'right',
+        '&:first-child': {
+          textAlign: 'left',
+        },
+      }}
+    >
+      {t(`discovery.table.header.${camelToKebab(label)}`)}
+    </Box>
+  )
+}
+
+export function DiscoveryTableDataRow({ row }: { row: DiscoveryTableRowData }) {
+  return (
+    <Box
+      as="tr"
+      sx={{
+        borderRadius: 'medium',
+        transition: 'box-shadow 200ms',
+        '&:hover': {
+          boxShadow: 'buttonMenu',
+        },
+      }}
+    >
+      {Object.keys(row).map((key) => (
+        <DiscoveryTableDataCell key={key} label={key} row={row} />
+      ))}
+    </Box>
+  )
+}
+
+export function DiscoveryTableDataCell({
+  label,
+  row,
+}: {
+  label: string
+  row: DiscoveryTableRowData
+}) {
+  return (
+    <Box
+      as="td"
+      sx={{
+        p: '8px 12px',
+        textAlign: 'right',
+        '&:first-child': {
+          textAlign: 'left',
+        },
+      }}
+    >
+      <DiscoveryTableDataCellContent label={label} row={row} />
     </Box>
   )
 }
