+++ features/discovery/discoveryApi.ts
@@ -8,7 +8,6 @@ import { ajax } from 'rxjs/ajax'
 import { catchError, map } from 'rxjs/operators'
 
 export interface DiscoveryDataResponse {
-  // TODO: replace with real discovery data when available
   data?: {
     rows: DiscoveryTableRowData[]
   }
